Share CV header parsing across metadata maintenance
